reorganizing input and output CSV files.  Some minor changes to the tasks.csv were made by hand:
I added the dependencies for the installation and commissioning work.  I also cut the duration of the installation work in half (a mistake generated in the original spreadsheet due to the fact that two people are assumed to be working in tandum). One should be careful not to overwrite these!